src/requirements-api.txt
# BID Web Backend — FastAPI dependencies
# Install from the project root: pip install -r src/requirements-api.txt

# ── Core web framework ────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ── Settings management ───────────────────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ── ORM / persistence ─────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.2.0

# ── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
email-validator>=2.0.0     # required by pydantic EmailStr

# ── Testing ───────────────────────────────────────────────────────────────────
httpx>=0.27.0          # async HTTP client used by FastAPI TestClient
pytest-asyncio>=0.23.0 # async test support
